+++ src/app/app.ts
@@ -1,8 +1,6 @@
 /**
  * Expressアプリケーション
- * @ignore
  */
-
 import * as sskts from '@motionpicture/sskts-domain';
 import * as bodyParser from 'body-parser';
 import * as createDebug from 'debug';
@@ -30,6 +28,7 @@ app.use(bodyParser.urlencoded({ extended: true }));
 // app.use(express.static(__dirname + '/../public'));
 
 // MongoDB接続
+// tslint:disable-next-line:no-console
 sskts.mongoose.connect(<string>process.env.MONGOLAB_URI, mongooseConnectionOptions).then(debug).catch(console.error);
 
 // routers