+++ src/mongooseConnectionOptions.ts
@@ -1,15 +1,17 @@
 /**
  * mongoose接続オプション
  * @see http://mongoosejs.com/docs/api.html#index_Mongoose-connect
- * @ignore
  */
-const mongooseConnectionOptions = {
+import { mongoose } from '@motionpicture/sskts-domain';
+
+const mongooseConnectionOptions: mongoose.ConnectionOptions = {
     autoReconnect: true,
     keepAlive: 120000,
     connectTimeoutMS: 30000,
     socketTimeoutMS: 0,
     reconnectTries: 30,
-    reconnectInterval: 1000
+    reconnectInterval: 1000,
+    useNewUrlParser: true
 };
 
 export default mongooseConnectionOptions;